extract the source code for session resumption and ticket-based resumption to dedicated files (so that we can add unit tests for them)
