+++ src/main.c
@@ -46,8 +46,6 @@
 #include <sys/wait.h>
 #include <openssl/crypto.h>
 #include <openssl/err.h>
-#include <openssl/evp.h>
-#include <openssl/rand.h>
 #include <openssl/ssl.h>
 #ifdef __GLIBC__
 #include <execinfo.h>
@@ -59,10 +57,10 @@
 #include "h2o/http1.h"
 #include "h2o/http2.h"
 #include "h2o/serverutil.h"
-
 #ifdef H2O_USE_MRUBY
 #include "h2o/mruby.h"
 #endif
+#include "standalone.h"
 
 /* simply use a large value, and let the kernel clip it to the internal max */
 #define H2O_SOMAXCONN (65535)
@@ -74,12 +72,6 @@
 #endif
 
 #define H2O_DEFAULT_NUM_NAME_RESOLUTION_THREADS 32
-
-#if defined(SSL_CTRL_SET_TLSEXT_TICKET_KEY_CB) && !defined(OPENSSL_NO_TLSEXT)
-#define H2O_USE_SESSION_TICKETS 1
-#else
-#define H2O_USE_SESSION_TICKETS 0
-#endif
 
 struct listener_ssl_config_t {
     H2O_VECTOR(h2o_iovec_t) hostnames;
@@ -120,16 +112,6 @@ typedef enum en_run_mode_t {
     RUN_MODE_TEST,
 } run_mode_t;
 
-struct st_session_ticket_internal_updater_conf_t {
-    const EVP_CIPHER *cipher;
-    const EVP_MD *md;
-    unsigned lifetime;
-};
-
-struct st_session_ticket_file_updater_conf_t {
-    const char *filename;
-};
-
 static struct {
     h2o_globalconf_t globalconf;
     run_mode_t run_mode;
@@ -146,23 +128,6 @@ static struct {
     int max_connections;
     size_t num_threads;
     int tfo_queues;
-    struct {
-        void *(*update_thread)(void *conf);
-        union {
-            struct st_session_ticket_internal_updater_conf_t internal;
-            struct st_session_ticket_file_updater_conf_t file;
-        } conf;
-    } session_ticket;
-    struct {
-        void (*setup)(void);
-        struct {
-            char *host;
-            uint16_t port;
-            size_t num_threads;
-            char *prefix;
-            unsigned timeout;
-        } memcached;
-    } session_resumption;
     struct {
         pthread_t tid;
         h2o_context_t ctx;
@@ -188,8 +153,6 @@ static struct {
     1024,            /* max_connections */
     0,               /* initialized in main() */
     0,               /* initialized in main() */
-    {},              /* session_tickets (initialized in main()) */
-    {},              /* memcached_session_resumption */
     NULL,            /* thread_ids */
     0,               /* shutdown_requested */
     {},              /* state */
@@ -229,20 +192,16 @@ static void openssl_thread_lock_callback(int mode, int n, const char *file, int 
 
 static void init_openssl(void)
 {
-    static int ready = 0;
-    if (!ready) {
-        int nlocks = CRYPTO_num_locks(), i;
-        openssl_thread_locks = h2o_mem_alloc(sizeof(*openssl_thread_locks) * nlocks);
-        for (i = 0; i != nlocks; ++i)
-            pthread_mutex_init(openssl_thread_locks + i, NULL);
-        CRYPTO_set_locking_callback(openssl_thread_lock_callback);
-        CRYPTO_set_id_callback(openssl_thread_id_callback);
-        /* TODO [OpenSSL] set dynlock callbacks for better performance */
-        SSL_load_error_strings();
-        SSL_library_init();
-        OpenSSL_add_all_algorithms();
-        ready = 1;
-    }
+    int nlocks = CRYPTO_num_locks(), i;
+    openssl_thread_locks = h2o_mem_alloc(sizeof(*openssl_thread_locks) * nlocks);
+    for (i = 0; i != nlocks; ++i)
+        pthread_mutex_init(openssl_thread_locks + i, NULL);
+    CRYPTO_set_locking_callback(openssl_thread_lock_callback);
+    CRYPTO_set_id_callback(openssl_thread_id_callback);
+    /* TODO [OpenSSL] set dynlock callbacks for better performance */
+    SSL_load_error_strings();
+    SSL_library_init();
+    OpenSSL_add_all_algorithms();
 }
 
 static void setup_ecc_key(SSL_CTX *ssl_ctx)
@@ -285,357 +244,6 @@ static int on_sni_callback(SSL *ssl, int *ad, void *arg)
 
     return SSL_TLSEXT_ERR_OK;
 }
-
-#if H2O_USE_SESSION_TICKETS
-
-struct st_session_ticket_t {
-    unsigned char name[16];
-    struct {
-        const EVP_CIPHER *cipher;
-        unsigned char *key;
-    } cipher;
-    struct {
-        const EVP_MD *md;
-        unsigned char *key;
-    } hmac;
-    uint64_t not_before;
-    uint64_t not_after;
-};
-
-typedef H2O_VECTOR(struct st_session_ticket_t *) session_ticket_vector_t;
-
-static struct {
-    pthread_rwlock_t rwlock;
-    session_ticket_vector_t tickets; /* sorted from newer to older */
-} session_tickets = {
-    PTHREAD_RWLOCK_INITIALIZER, /* rwlock (FIXME need to favor writer over readers explicitly, the default of Linux is known to be
-                                   the otherwise) */
-    {}                          /* tickets */
-};
-
-static int session_ticket_key_callback(SSL *ssl, unsigned char *key_name, unsigned char *iv, EVP_CIPHER_CTX *ctx, HMAC_CTX *hctx,
-                                       int enc)
-{
-    int ret;
-    pthread_rwlock_rdlock(&session_tickets.rwlock);
-
-    if (enc) {
-        if (session_tickets.tickets.size == 0) {
-            ret = -1;
-            goto Exit;
-        }
-        struct st_session_ticket_t *ticket = session_tickets.tickets.entries[0];
-        memcpy(key_name, ticket->name, sizeof(ticket->name));
-        RAND_pseudo_bytes(iv, EVP_MAX_IV_LENGTH);
-        EVP_EncryptInit_ex(ctx, ticket->cipher.cipher, NULL, ticket->cipher.key, NULL);
-        HMAC_Init_ex(hctx, ticket->hmac.key, ticket->hmac.md->block_size, ticket->hmac.md, NULL);
-        ret = 1;
-    } else {
-        struct st_session_ticket_t *ticket;
-        size_t i;
-        for (i = 0; i != session_tickets.tickets.size; ++i) {
-            ticket = session_tickets.tickets.entries[i];
-            if (memcmp(ticket->name, key_name, sizeof(ticket->name)) == 0)
-                goto Found;
-        }
-        /* not found */
-        ret = 0;
-        goto Exit;
-    Found:
-        EVP_DecryptInit_ex(ctx, ticket->cipher.cipher, NULL, ticket->cipher.key, NULL);
-        HMAC_Init_ex(hctx, ticket->hmac.key, ticket->hmac.md->block_size, ticket->hmac.md, NULL);
-        ret = i == 0 ? 1 : 2; /* request renew if the key is not the newest one */
-    }
-
-Exit:
-    pthread_rwlock_unlock(&session_tickets.rwlock);
-    return ret;
-}
-
-struct st_session_ticket_t *session_ticket_new(const EVP_CIPHER *cipher, const EVP_MD *md, uint64_t not_before, uint64_t not_after,
-                                               int fill_in)
-{
-    struct st_session_ticket_t *ticket = h2o_mem_alloc(sizeof(*ticket) + cipher->key_len + md->block_size);
-
-    ticket->cipher.cipher = cipher;
-    ticket->cipher.key = (unsigned char *)ticket + sizeof(*ticket);
-    ticket->hmac.md = md;
-    ticket->hmac.key = ticket->cipher.key + cipher->key_len;
-    ticket->not_before = not_before;
-    ticket->not_after = not_after;
-    if (fill_in) {
-        RAND_pseudo_bytes(ticket->name, sizeof(ticket->name));
-        RAND_pseudo_bytes(ticket->cipher.key, ticket->cipher.cipher->key_len);
-        RAND_pseudo_bytes(ticket->hmac.key, ticket->hmac.md->block_size);
-    }
-
-    return ticket;
-}
-
-static void session_ticket_free(struct st_session_ticket_t *ticket)
-{
-    h2o_mem_set_secure(ticket, 0, sizeof(*ticket) + ticket->cipher.cipher->key_len + ticket->hmac.md->block_size);
-    free(ticket);
-}
-
-static int session_ticket_sort_compare(const void *_x, const void *_y)
-{
-    struct st_session_ticket_t *x = *(void **)_x, *y = *(void **)_y;
-
-    if (x->not_before != y->not_before)
-        return x->not_before > y->not_before ? -1 : 1;
-    return memcmp(x->name, y->name, sizeof(x->name));
-}
-
-static void session_ticket_free_vector(session_ticket_vector_t *tickets)
-{
-    size_t i;
-    for (i = 0; i != tickets->size; ++i)
-        session_ticket_free(tickets->entries[i]);
-    free(tickets->entries);
-    memset(tickets, 0, sizeof(*tickets));
-}
-
-static void *session_ticket_internal_updater(void *_conf)
-{
-    struct st_session_ticket_internal_updater_conf_t *conf = _conf;
-
-    while (1) {
-        uint64_t newest_not_before = 0, oldest_not_after = UINT64_MAX, now = time(NULL);
-
-        /* obtain modification times */
-        pthread_rwlock_rdlock(&session_tickets.rwlock);
-        if (session_tickets.tickets.size != 0) {
-            newest_not_before = session_tickets.tickets.entries[0]->not_before;
-            oldest_not_after = session_tickets.tickets.entries[session_tickets.tickets.size - 1]->not_after;
-        }
-        pthread_rwlock_unlock(&session_tickets.rwlock);
-
-        /* insert new entry if necessary */
-        if (newest_not_before + conf->lifetime / 4 <= now) {
-            struct st_session_ticket_t *new_ticket = session_ticket_new(conf->cipher, conf->md, now, now + conf->lifetime - 1, 1);
-            pthread_rwlock_wrlock(&session_tickets.rwlock);
-            h2o_vector_reserve(NULL, (void *)&session_tickets.tickets, sizeof(session_tickets.tickets.entries[0]),
-                               session_tickets.tickets.size + 1);
-            memmove(session_tickets.tickets.entries + 1, session_tickets.tickets.entries,
-                    sizeof(session_tickets.tickets.entries[0]) * session_tickets.tickets.size);
-            session_tickets.tickets.entries[0] = new_ticket;
-            ++session_tickets.tickets.size;
-            pthread_rwlock_unlock(&session_tickets.rwlock);
-        }
-
-        /* free expired entries if necessary */
-        if (oldest_not_after < now) {
-            while (1) {
-                struct st_session_ticket_t *expiring_ticket = NULL;
-                pthread_rwlock_wrlock(&session_tickets.rwlock);
-                if (session_tickets.tickets.size != 0 &&
-                    session_tickets.tickets.entries[session_tickets.tickets.size - 1]->not_after < now) {
-                    expiring_ticket = session_tickets.tickets.entries[session_tickets.tickets.size - 1];
-                    session_tickets.tickets.entries[session_tickets.tickets.size - 1] = NULL;
-                    --session_tickets.tickets.size;
-                }
-                pthread_rwlock_unlock(&session_tickets.rwlock);
-                if (expiring_ticket == NULL)
-                    break;
-                session_ticket_free(expiring_ticket);
-            }
-        }
-
-        /* sleep for certain amount of time */
-        sleep(120 - (rand() >> 16) % 7);
-    }
-}
-
-static void session_ticket_internal_updater_set_defaults(void)
-{
-    init_openssl();
-    /* to protect the secret >>>2030 we need AES-256 (http://www.keylength.com/en/4/), sha1 is used only during the communication
-     * and can be short */
-    conf.session_ticket.update_thread = session_ticket_internal_updater;
-    conf.session_ticket.conf.internal.cipher = EVP_aes_256_cbc();
-    conf.session_ticket.conf.internal.md = EVP_sha1();
-    conf.session_ticket.conf.internal.lifetime = 3600; /* 1 hour */
-}
-
-static struct st_session_ticket_t *session_ticket_parse_element(yoml_t *element, char *errstr)
-{
-    yoml_t *t;
-    struct st_session_ticket_t *ticket;
-    unsigned char name[sizeof(ticket->name) + 1], *key;
-    const EVP_CIPHER *cipher;
-    const EVP_MD *hash;
-    uint64_t not_before, not_after;
-
-    errstr[0] = '\0';
-
-    if (element->type != YOML_TYPE_MAPPING) {
-        strcpy(errstr, "node is not a mapping");
-        return NULL;
-    }
-
-#define FETCH(n, post)                                                                                                             \
-    do {                                                                                                                           \
-        if ((t = yoml_get(element, n)) == NULL) {                                                                                  \
-            strcpy(errstr, " mandatory attribute `" n "` is missing");                                                             \
-            return NULL;                                                                                                           \
-        }                                                                                                                          \
-        if (t->type != YOML_TYPE_SCALAR) {                                                                                         \
-            strcpy(errstr, "attribute `" n "` is not a string");                                                                   \
-            return NULL;                                                                                                           \
-        }                                                                                                                          \
-        post                                                                                                                       \
-    } while (0)
-
-    FETCH("name", {
-        if (strlen(t->data.scalar) != sizeof(ticket->name) * 2) {
-            strcpy(errstr, "length of `name` attribute is not 32 bytes");
-            return NULL;
-        }
-        if (h2o_hex_decode(name, t->data.scalar, sizeof(ticket->name) * 2) != 0) {
-            strcpy(errstr, "failed to decode the hex-encoded name");
-            return NULL;
-        }
-    });
-    FETCH("cipher", {
-        if ((cipher = EVP_get_cipherbyname(t->data.scalar)) == NULL) {
-            strcpy(errstr, "cannot find the named cipher algorithm");
-            return NULL;
-        }
-    });
-    FETCH("hash", {
-        if ((hash = EVP_get_digestbyname(t->data.scalar)) == NULL) {
-            strcpy(errstr, "cannot find the named hash algorgithm");
-            return NULL;
-        }
-    });
-    FETCH("key", {
-        size_t keylen = cipher->key_len + hash->block_size;
-        if (strlen(t->data.scalar) != keylen * 2) {
-            sprintf(errstr, "length of the `key` attribute is incorrect (is %zu, must be %zu)\n", strlen(t->data.scalar),
-                    keylen * 2);
-            return NULL;
-        }
-        key = alloca(keylen + 1);
-        if (h2o_hex_decode(key, t->data.scalar, keylen * 2) != 0) {
-            strcpy(errstr, "failed to decode the hex-encoded key");
-            return NULL;
-        }
-    });
-    FETCH("not_before", {
-        if (sscanf(t->data.scalar, "%" SCNu64, &not_before) != 1) {
-            strcpy(errstr, "failed to parse the `not_before` attribute");
-            return NULL;
-        }
-    });
-    FETCH("not_after", {
-        if (sscanf(t->data.scalar, "%" SCNu64, &not_after) != 1) {
-            strcpy(errstr, "failed to parse the `not_after` attribute");
-            return NULL;
-        }
-    });
-    if (!(not_before <= not_after)) {
-        strcpy(errstr, "`not_after` is not equal to or greater than `not_before`");
-        return NULL;
-    }
-
-#undef FETCH
-
-    ticket = session_ticket_new(cipher, hash, not_before, not_after, 0);
-    memcpy(ticket->name, name, sizeof(ticket->name));
-    memcpy(ticket->cipher.key, key, cipher->key_len);
-    memcpy(ticket->hmac.key, key + cipher->key_len, hash->block_size);
-    return ticket;
-}
-
-static int session_ticket_file_updater_load_file(struct st_session_ticket_file_updater_conf_t *conf)
-{
-#define ERR_PREFIX "failed to load session ticket secrets from file:%s:"
-
-    yoml_t *yoml = NULL;
-    FILE *fp;
-    yaml_parser_t parser;
-    session_ticket_vector_t tickets = {};
-    size_t i;
-    int ret = -1;
-
-    yaml_parser_initialize(&parser);
-
-    /* load yaml */
-    if ((fp = fopen(conf->filename, "rb")) == NULL) {
-        char errbuf[256];
-        strerror_r(errno, errbuf, sizeof(errbuf));
-        fprintf(stderr, ERR_PREFIX "%s\n", conf->filename, errbuf);
-        goto Exit;
-    }
-    yaml_parser_set_input_file(&parser, fp);
-    if ((yoml = yoml_parse_document(&parser, NULL, conf->filename)) == NULL) {
-        fprintf(stderr, ERR_PREFIX "parse error at line %d:%s\n", conf->filename, (int)parser.problem_mark.line, parser.problem);
-        goto Exit;
-    }
-    /* parse the data */
-    if (yoml->type != YOML_TYPE_SEQUENCE) {
-        fprintf(stderr, ERR_PREFIX "root element is not a sequence\n", conf->filename);
-        goto Exit;
-    }
-    for (i = 0; i != yoml->data.sequence.size; ++i) {
-        char errbuf[256];
-        struct st_session_ticket_t *ticket = session_ticket_parse_element(yoml->data.sequence.elements[i], errbuf);
-        if (ticket == NULL) {
-            fprintf(stderr, ERR_PREFIX "at element index %zu:%s\n", conf->filename, i, errbuf);
-            goto Exit;
-        }
-        h2o_vector_reserve(NULL, (void *)&tickets, sizeof(tickets.entries[0]), tickets.size + 1);
-        tickets.entries[tickets.size++] = ticket;
-    }
-    /* sort the ticket entries being read */
-    qsort(tickets.entries, tickets.size, sizeof(tickets.entries[0]), session_ticket_sort_compare);
-    /* replace the ticket list */
-    pthread_rwlock_wrlock(&session_tickets.rwlock);
-    session_ticket_free_vector(&session_tickets.tickets);
-    session_tickets.tickets = tickets;
-    pthread_rwlock_unlock(&session_tickets.rwlock);
-    tickets = (session_ticket_vector_t){};
-
-    ret = 0;
-
-Exit:
-    if (fp != NULL)
-        fclose(fp);
-    yaml_parser_delete(&parser);
-    if (yoml != NULL)
-        yoml_free(yoml);
-    session_ticket_free_vector(&tickets);
-    return ret;
-
-#undef ERR_PREFIX
-}
-
-static void *session_ticket_file_updater(void *_conf)
-{
-    struct st_session_ticket_file_updater_conf_t *conf = _conf;
-    time_t last_mtime = 1; /* file is loaded if mtime changes, 0 is used to indicate that the file was missing */
-
-    while (1) {
-        struct stat st;
-        if (stat(conf->filename, &st) != 0) {
-            if (last_mtime != 0) {
-                char errbuf[256];
-                strerror_r(errno, errbuf, sizeof(errbuf));
-                fprintf(stderr, "cannot load session ticket secrets from file:%s:%s\n", conf->filename, errbuf);
-            }
-            last_mtime = 0;
-        } else if (last_mtime != st.st_mtime) {
-            /* (re)load */
-            last_mtime = st.st_mtime;
-            if (session_ticket_file_updater_load_file(conf) == 0)
-                fprintf(stderr, "session ticket secrets have been (re)loaded\n");
-        }
-        sleep(10);
-    }
-}
-
-#endif
 
 #ifndef OPENSSL_NO_OCSP
 
@@ -900,7 +508,6 @@ static int listener_setup_ssl(h2o_configurator_command_t *cmd, h2o_configurator_
 #endif
 
     /* setup */
-    init_openssl();
     ssl_ctx = SSL_CTX_new(SSLv23_server_method());
     SSL_CTX_set_options(ssl_ctx, ssl_options);
 
@@ -1416,183 +1023,6 @@ static int on_config_num_threads(h2o_configurator_command_t *cmd, h2o_configurat
     return 0;
 }
 
-static int on_config_ssl_session_ticket(h2o_configurator_command_t *cmd, h2o_configurator_context_t *ctx, yoml_t *node)
-{
-    yoml_t *t;
-
-    if ((t = yoml_get(node, "mode")) == NULL) {
-        h2o_configurator_errprintf(cmd, node, "mandatory attribute `mode` is missing");
-        return -1;
-    }
-
-    if (t->type == YOML_TYPE_SCALAR) {
-
-        if (strcasecmp(t->data.scalar, "off") == 0) {
-
-            /* mode: off */
-            conf.session_ticket.update_thread = NULL;
-            return 0;
-
-        } else if (strcasecmp(t->data.scalar, "internal") == 0) {
-
-            /* mode: internal takes three arguments: cipher, hash, duration */
-            session_ticket_internal_updater_set_defaults();
-            if ((t = yoml_get(node, "cipher")) != NULL) {
-                if (t->type != YOML_TYPE_SCALAR ||
-                    (conf.session_ticket.conf.internal.cipher = EVP_get_cipherbyname(t->data.scalar)) == NULL) {
-                    h2o_configurator_errprintf(cmd, t, "unknown cipher algorithm");
-                    return -1;
-                }
-            }
-            if ((t = yoml_get(node, "hash")) != NULL) {
-                if (t->type != YOML_TYPE_SCALAR ||
-                    (conf.session_ticket.conf.internal.md = EVP_get_digestbyname(t->data.scalar)) == NULL) {
-                    h2o_configurator_errprintf(cmd, t, "unknown hash algorithm");
-                    return -1;
-                }
-            }
-            if ((t = yoml_get(node, "lifetime")) != NULL) {
-                if (t->type != YOML_TYPE_SCALAR || sscanf(t->data.scalar, "%u", &conf.session_ticket.conf.internal.lifetime) != 1 ||
-                    conf.session_ticket.conf.internal.lifetime == 0) {
-                    h2o_configurator_errprintf(cmd, t, "`liftime` must be a positive number (in seconds)");
-                    return -1;
-                }
-            }
-            return 0;
-
-        } else if (strcasecmp(t->data.scalar, "file") == 0) {
-
-            /* mode: file reads the contents of the file and uses it as the session ticket secret */
-            init_openssl();
-            if ((t = yoml_get(node, "file")) == NULL) {
-                h2o_configurator_errprintf(cmd, node, "mandatory attribute `file` is missing");
-                return -1;
-            }
-            if (t->type != YOML_TYPE_SCALAR) {
-                h2o_configurator_errprintf(cmd, node, "`file` must be a string");
-                return -1;
-            }
-            conf.session_ticket.update_thread = session_ticket_file_updater;
-            conf.session_ticket.conf.file.filename = h2o_strdup(NULL, t->data.scalar, SIZE_MAX).base;
-            return 0;
-        }
-    }
-
-    h2o_configurator_errprintf(cmd, t, "`mode` must be one of: `internal`, `off`");
-    return -1;
-}
-
-static void setup_session_resumption_disable(void)
-{
-    size_t i, j;
-    for (i = 0; i != conf.num_listeners; ++i) {
-        for (j = 0; j != conf.listeners[i]->ssl.size; ++j) {
-            SSL_CTX *ctx = conf.listeners[i]->ssl.entries[j]->ctx;
-            SSL_CTX_set_session_cache_mode(ctx, SSL_SESS_CACHE_OFF);
-        }
-    }
-}
-
-static void setup_session_resumption_memcached(void)
-{
-    h2o_memcached_context_t *memc_ctx =
-        h2o_memcached_create_context(conf.session_resumption.memcached.host, conf.session_resumption.memcached.port,
-                                     conf.session_resumption.memcached.num_threads, conf.session_resumption.memcached.prefix);
-    h2o_accept_setup_async_ssl_resumption(memc_ctx, conf.session_resumption.memcached.timeout);
-    size_t i;
-    for (i = 0; i != conf.num_listeners; ++i)
-        if (conf.listeners[i]->ssl.size != 0)
-            h2o_socket_ssl_async_resumption_setup_ctx(conf.listeners[i]->ssl.entries[0]->ctx);
-}
-
-static int on_config_ssl_session_resumption(h2o_configurator_command_t *cmd, h2o_configurator_context_t *ctx, yoml_t *node)
-{
-    yoml_t *mode;
-
-    if ((mode = yoml_get(node, "mode")) == NULL) {
-        h2o_configurator_errprintf(cmd, mode, "mandatory attribute `mode` is missing");
-        return -1;
-    }
-
-    if (mode->type == YOML_TYPE_SCALAR) {
-
-        if (strcmp(mode->data.scalar, "off") == 0) {
-
-            conf.session_resumption.setup = setup_session_resumption_disable;
-            return 0;
-
-        } else if (strcmp(mode->data.scalar, "internal") == 0) {
-
-            conf.session_resumption.setup = NULL;
-            return 0;
-
-        } else if (strcmp(mode->data.scalar, "memcached") == 0) {
-
-            const char *host = NULL, *prefix = ":h2o:ssl-resumption:";
-            uint16_t port = 11211;
-            size_t num_threads = 1;
-            unsigned timeout = 3600;
-            size_t index;
-            for (index = 0; index != node->data.mapping.size; ++index) {
-                yoml_t *key = node->data.mapping.elements[index].key;
-                yoml_t *value = node->data.mapping.elements[index].value;
-                if (value == mode)
-                    continue;
-                if (key->type != YOML_TYPE_SCALAR) {
-                    h2o_configurator_errprintf(cmd, key, "attribute must be a string");
-                    return -1;
-                }
-                if (strcmp(key->data.scalar, "host") == 0) {
-                    if (value->type != YOML_TYPE_SCALAR) {
-                        h2o_configurator_errprintf(cmd, value, "`host` must be a string");
-                        return -1;
-                    }
-                    host = value->data.scalar;
-                } else if (strcmp(key->data.scalar, "port") == 0) {
-                    if (!(value->type == YOML_TYPE_SCALAR && sscanf(value->data.scalar, "%" SCNu16, &port) == 1)) {
-                        h2o_configurator_errprintf(cmd, value, "`port` must be a number");
-                        return -1;
-                    }
-                } else if (strcmp(key->data.scalar, "num-threads") == 0) {
-                    if (!(value->type == YOML_TYPE_SCALAR && sscanf(value->data.scalar, "%zu", &num_threads) == 1 &&
-                          num_threads != 0)) {
-                        h2o_configurator_errprintf(cmd, value, "`num-threads` must be a positive number");
-                        return -1;
-                    }
-                } else if (strcmp(key->data.scalar, "prefix") == 0) {
-                    if (value->type != YOML_TYPE_SCALAR) {
-                        h2o_configurator_errprintf(cmd, value, "`prefix` must be a string");
-                        return -1;
-                    }
-                    prefix = value->data.scalar;
-                } else if (strcmp(key->data.scalar, "timeout") == 0) {
-                    if (!(value->type == YOML_TYPE_SCALAR && sscanf(value->data.scalar, "%u", &timeout) == 1 && timeout != 0)) {
-                        h2o_configurator_errprintf(cmd, value, "`timeout` must be a positive number (in seconds)");
-                        return -1;
-                    }
-                } else {
-                    h2o_configurator_errprintf(cmd, key, "unknown attribute: %s", key->data.scalar);
-                    return -1;
-                }
-            }
-            if (host == NULL) {
-                h2o_configurator_errprintf(cmd, node, "mandatory attribute `host` is missing");
-                return -1;
-            }
-            conf.session_resumption.setup = setup_session_resumption_memcached;
-            conf.session_resumption.memcached.host = h2o_strdup(NULL, host, SIZE_MAX).base;
-            conf.session_resumption.memcached.port = port;
-            conf.session_resumption.memcached.num_threads = num_threads;
-            conf.session_resumption.memcached.prefix = h2o_strdup(NULL, prefix, SIZE_MAX).base;
-            conf.session_resumption.memcached.timeout = timeout;
-            return 0;
-        }
-    }
-
-    h2o_configurator_errprintf(cmd, mode, "`mode` must be one of: `off`, `internal`, `memached`");
-    return -1;
-}
-
 static int on_config_num_name_resolution_threads(h2o_configurator_command_t *cmd, h2o_configurator_context_t *ctx, yoml_t *node)
 {
     if (h2o_configurator_scanf(cmd, node, "%zu", &h2o_hostinfo_max_threads) != 0)
@@ -1968,10 +1398,10 @@ static void setup_configurators(void)
         h2o_configurator_define_command(c, "tcp-fastopen", H2O_CONFIGURATOR_FLAG_GLOBAL, on_config_tcp_fastopen);
         h2o_configurator_define_command(c, "ssl-session-ticket",
                                         H2O_CONFIGURATOR_FLAG_GLOBAL | H2O_CONFIGURATOR_FLAG_EXPECT_MAPPING,
-                                        on_config_ssl_session_ticket);
+                                        ssl_session_ticket_on_config);
         h2o_configurator_define_command(c, "ssl-session-resumption",
                                         H2O_CONFIGURATOR_FLAG_GLOBAL | H2O_CONFIGURATOR_FLAG_EXPECT_MAPPING,
-                                        on_config_ssl_session_resumption);
+                                        ssl_session_resumption_on_config);
     }
 
     h2o_access_log_register_configurator(&conf.globalconf);
@@ -1995,11 +1425,10 @@ int main(int argc, char **argv)
     conf.num_threads = h2o_numproc();
     conf.tfo_queues = H2O_DEFAULT_LENGTH_TCP_FASTOPEN_QUEUE;
 
-#if H2O_USE_SESSION_TICKETS
-    session_ticket_internal_updater_set_defaults();
-#endif
-
     h2o_hostinfo_max_threads = H2O_DEFAULT_NUM_NAME_RESOLUTION_THREADS;
+
+    init_openssl();
+    ssl_session_ticket_init();
     setup_configurators();
 
     { /* parse options */
@@ -2204,38 +1633,20 @@ int main(int argc, char **argv)
         fclose(fp);
     }
 
-    /* start memcached client threads for session resumption, and adjust the SSL contexts */
-    if (conf.session_resumption.setup != NULL)
-        conf.session_resumption.setup();
-
-#if H2O_USE_SESSION_TICKETS
-    /* start session ticket updater thread */
-    if (conf.session_ticket.update_thread != NULL) {
+    { /* initialize SSL_CTXs for session resumption and ticket-based resumption (also starts memcached client threads for the
+         purpose) */
         size_t i, j;
-        int sslfound = 0;
+        H2O_VECTOR(SSL_CTX *)ssl_contexts = {};
         for (i = 0; i != conf.num_listeners; ++i) {
             for (j = 0; j != conf.listeners[i]->ssl.size; ++j) {
-                SSL_CTX_set_tlsext_ticket_key_cb(conf.listeners[i]->ssl.entries[j]->ctx, session_ticket_key_callback);
-                sslfound = 1;
+                h2o_vector_reserve(NULL, (void *)&ssl_contexts, sizeof(ssl_contexts.entries[0]), ssl_contexts.size + 1);
+                ssl_contexts.entries[ssl_contexts.size++] = conf.listeners[i]->ssl.entries[j]->ctx;
             }
         }
-        if (sslfound) {
-            pthread_t tid;
-            pthread_attr_t attr;
-            pthread_attr_init(&attr);
-            pthread_attr_setdetachstate(&attr, 1);
-            h2o_multithread_create_thread(&tid, &attr, conf.session_ticket.update_thread, &conf.session_ticket.conf);
-        }
-    } else {
-        size_t i, j;
-        for (i = 0; i != conf.num_listeners; ++i) {
-            for (j = 0; j != conf.listeners[i]->ssl.size; ++j) {
-                SSL_CTX *ctx = conf.listeners[i]->ssl.entries[j]->ctx;
-                SSL_CTX_set_options(ctx, SSL_CTX_get_options(ctx) | SSL_OP_NO_TICKET);
-            }
-        }
+        ssl_session_resumption_setup(ssl_contexts.entries, ssl_contexts.size);
+        ssl_session_ticket_setup(ssl_contexts.entries, ssl_contexts.size);
+        free(ssl_contexts.entries);
     }
-#endif
 
     /* all setup should be complete by now */
